core: Add Thing.Junk method to remove items from the world
This commit adds the Thing.Junk method to remove items from the world.
If an item is unique it will be moved back to it's origin and a reset
scheduled. If an item is spawnable it will be scheduled for a reset. If
the item is not unique or spawnable it will be freed for garbage
collection. Inventories of junked items are also junked recursively.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>