core: Add Thing.Is constant Freed
This commit adds a Thing.Is constant of Freed to flag when a Thing has
been freed for garbage collection. This is not technically required as
Thing.Free can be called multiple times without issue. However it is
useful for tracking and debugging.

The Thing.Free method has been updated to check and set Thing.Is[Freed].

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>