core: Fix sequence of recursive junking in Thing.Junk
Normally it is okay to either process In then Out for a Thing or Out
then In - the sequence does not matter. However, in Thing.Junk it is
possible for an item to be scheduled for a reset - In which case the
item will be moved from In to Out. In Junk if In is processed before Out
an item may be processed twice - once when found in In and move to Out
and then again when Out is processed. It is therefore important to
process Out before In to prevent this,

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>