core: Thing.Free not clearing some items, add note to Thing type
In Thing.Free some of the Thing fields are not cleared and some not
cleared very well. This commit improves the Free command to cover all
current Thing fields. A note has also been added to the Thing type about
also updating NewThing and Thing.Free methods when adding fields.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>