core: Add clean-up handling to DROP, GET, PUT and TAKE commands
This commit schedules clean-up events for the DROP and PUT commands and
cancels clean-up events for GET and TAKE commands.

When an item is dropped a clean-up event is scheduled.

When an item is taken from a container we cancel any pending clean-up
events.

When an item is put into a container a clean-up event is only registered
if the container is not help by a player and does not have a pending
clean-up event. If a container has a pending clean-up event then the
items in it will get cleaned-up along with the container. If the
container is help by a player then we don't want their items just
disappearing!

This allows items to be put into an 'original' container, one that has
not been spawned yet, and still have the items cleaned-up so they do not
accumulate.

When an item is picked up any clean-up event is cancelled. If it's a
container then any clean-up events of immediate inventory items are
cancelled. This cancels clean-up events for items put into 'original'
containers.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>