core: Add Trigger event definitions and handler
This commit adds the definitions and handler for Trigger events. A
Trigger event works like other events such as Action, Cleanup and Reset.
However it is more general and intended to be used internally - rather
than specifically driven from zone files. For example a trigger could be
used to automatically open/close a door, extinguish a spent light or
fire, expire a poison or other applied effect.

To use a trigger event the item should have Thing.As[TriggerType] set to
a string ID used to identify the type of trigger. Then Schedule(Trigger)
cam be called. When the Trigger event fires the $TRIGGER handler will be
called - which should have a case statement setup for the specific
trigger ID being used - otherwise nothing happens.

For example trigger see the BLOCKER trigger used by Blockers/Doors and
added in a separate commit.

Note: The trigger handler implementation is not ideal. Maybe we could
register functions like the events themselves. For now this works...

Note: At the moment we only support one TriggerType per Thing. However,
TriggerType could be made a Thing.Any instead of a Thing.Is field.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>