core: Unmarshal RESET+JITTER attr for DOOR field + setup Trigger
Currently doors do not automatically return to their initial open/closed
state. This commit adds unmarshaling for the RESET and JITTER attributes
for the DOOR field and stores it in the Thing. The RESET and JITTER are
used to setup a Trigger event and a TriggerType of BLOCKER is set. The
initial state of the door, open or closed, is stored in Thing.Is as
_Open. This provides all of the required information so that the OPEN
and CLOSE commands can be updated to schedule the Trigger event - to be
completed in a separate commit.

The Thing.Is constants have been updated to add _Open.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>