core: Add BLOCKER Trigger handling + update OPEN/CLOSE
This commit updates the OPEN and CLOSE commands to schedule a BLOCKER
Trigger event to return a blocker/door to its initial open/closed state
after a period of time.

The trigger will only have effect if RESET and/or JITTER have been
defined for the DOOR field. Otherwise the door is just a manually
operated door.

The Trigger function has been updated to implement handling of the
BLOCKER trigger type.

The OPEN and CLOSE commands have also been updated so that the messages
make sense when it is the blocker/door opening or closing itself. Crowd
control has also been improved.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>