core: Update Thing.Unmarshal for Body field
This commit updates the Thing.Unmarshal method to unmarshal Body record
jar fields into Thing.Any[Body]. A new anyKey constant has been added
for Body.

The Body field contains the body slots available for holding, wearing or
wielding items. For example a simple Body may be defined as:

  Body: HEAD SHOULDERS CHEST ARM→2 HAND→2 WAIST LEG→2 FOOT→2

When unmarshaled the individual slots are represented in Thing.As[Body]:

  Thing.Any[Body] = []string{"HEAD", SHOULDERS", "CHEST", "ARM", "ARM",
    "HAND", "HAND", "WAIST", "LEG", "LEG", "FOOT", "FOOT"}

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>