core: Update Thing.Unmarshal for Holdable, Wearable and Wieldable fields
This commit updates the Thing.Unmarshal method to unmarshal Holdable,
Wearable and Wieldable record jar fields into Thing.Any[Holdable],
Thing.Any[Wearable] and Thing.Any[Wieldable]. New anyKey constants have
been added for Holdable, Wearable and Wieldable.

The Holdable, Wearable and Wieldable fields contain the body slots
required to either hold, wear or wield an item. For example a large
battle axe that requires two hands to wield and one hand to hold might
be defined as having:

  Wieldable: HAND→2
   Holdable: HAND

When unmarshaled the individual slots are represented in the Thing.As
values as in:

  Thing.Any[Wieldable] = []string{"HAND", "HAND"}
  Thing.Any[Holdable]  = []string{"HAND"}

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>