core: In Thing.Spawn clear spawnable flag properly
In Thing.Spawn the Spawnable bit in Thing.Is is cleared using ^= (XOR)
which actually toggles the bit. In this particular case this happens to
work because we only spawn items with the Spawnable bit set. However,
the Spawnable bit should more correctly be cleared using &^= (AND NOT).

This commit corrects the clearing of the Spanable bit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>