core: Add simple body check to identify NPCs
Currently mobiles/NPCs are not identified as such, they are just seen as
additional items in the game world. However, mobiles/NPCs should be
identified and flagged so that commands and interactions can be tailored
in much the same way as players are handled now.

This commit updates Thing.Unmarshal to set the NPC flag in Thing.Is for
any item that has a Thing.Any[Body] and is not a player.

This is not an ideal solution - we currently have the frog and rabbit
for example with no body defined. In the future a specific mobile/NPC
field will probably be added to the zone files.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>