core: Enable most items to be held by default
This commit automatically enables items that are not locations, players,
NPCs or narratives to be held. In Thing.Unmarshal a "HAND" is added to
t.Any[Holdable] if nothing has been explicitly specified and the item is
not a location, player, NPC or narrative.

For automatically holdable items the HOLD command can be vetoed. For
items that are not normally holdable the Holdable recordjar field can be
explicitly defined - this also allows items to require two hands in
order to be held by defining "Holdable: HAND→2".

It would be preferable to perform the checks in state.Hold and not add
"HAND to all Thing.Any[Holdable]. However the slots are used for body
slot accounting and dynamically adding and removing slots would increase
garbage collection at the expense of a negligible memory reduction. This
method is also simpler than dynamically adding and removing the slots.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>