core: Reset usage in QUIT command when force junking
When a player QUITs their inventory is forcibly junked. However if a
unique item is being held, worn or wielded it will be reset without the
appropriate flags being cleared.

This commit clears the Holding, Wearing and Wielding flags of an item
before forcibly junking it.

As noted in the FIXME comment, there may be a more suitable place to add
the code than the QUIT command so that the flags are always cleared.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>