core: Allow GET on holdable Players/NPCs
This commit updates the GET command so that players or NPCs that are
defined as being holdable can be picked up. For example a rabbit could
be defined as being holdable in two hands "Holdable: HAND→2" in which
case it should be possible to pick it up. We should not really be
allowing holdable players though - maybe this should be tightened up?

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>