core: Players and NPCs are also not containers
This commit updates the check in Thing.Unmarshal for items that are not
containers to include Players and NPCs. We shouldn't be able to put
items into and take items from players and NPCs. This also fixes an
issue with the EXAMINE command where the inventory of NPCs is being
described like container content.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>