core: Rename Thing.Enable→Thing.InitOnce
This commit renames the Thing.Enable method to Thing.InitOnce to better
reflest its purpose. We also need to be able to split initialisation
into two parts. The first part is initialisation that is only performed
once when things are initially loaded and put into the world. The second
part is repeatable initialisation when items are put into the world or
reset.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>