core: Load initially held, worn and/or wielded items
This commit updates Thing.Unmarshal to load the Holding, Wearing and
Wielding fields from zone files. These fields are stored as
Thing.Any[_Holding], Thing.Any[_Wearing] and Thing.Any[_Wielding]. The
underscore indicates an initial value - as also used for the _Open value
used to store the initial state of a blocker/door.

The Holding, Wearing and Wielding fields contain references to items.
Therefore this commit also updates the Thing.InitOnce method to rewrite
the references to UIDs. This requires Thing.Unmarshal to fully qualify
the references with the zone.

The anyKey constants _Holding, _Wearing and _Wielding have been added to
core/types.go and the comments for Holding, Wearing and Wielding updated
for clarity.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>