core: Simplify Thing.Unmarshal by quilifying refs during not after
This commit simplifies Thing.Unmarshal by qualifying any references with
the zone information while unmarshaling the values, instead of iterating
over the values at the end.

The world.Load function now adds a colon ':' suffix to the zone in
Thing.As[Zone] to avoid repetitive concatenation of the colon in
Thing.Unmarshal.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>