core,cmd: Should be writing CRLF line endings to the network
When writing data to clients over the network we should be using CRLF
network line endings. This does not seem to effect TELNET on Linux but
totally messes up the display when using the standard TELNET client on
Windows.

There are three possible solutions:

  1. Call bytes.ReplaceAll to rewrite "\n" as "\r\n" when writing the
     data to the network in the cmd/server/main.player function. This
     would process every message to every client as it is sent.

  2. Call strings.ReplaceAll to rewrite "\n" as "\r\n" when writing to
     the message buffer in the core/state.Msg and core/state.MsgAppend
     methods. This would only process each message once independent of
     the number of clients it is sent to.

  3. Compose the data using "\r\n" instead of "\n" in the first place,
     then there is no need for any conversions.

This commit implements solution 3. The only overhead is when the server
is initially started and zone files are read - the Description field is
rewritten to use CRLF instead of LF.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>