cmd,core: Add a little colour to command messages
This commit adds colours to messages sent to the players. In general
messages use text.Good, text.Info and text.Bad instead of text.Green,
text.Yellow and text.Red as this allows the scheme to be easily changed.

This commit also updates the server.player function to reset to default
colours before displaying a new message. The function also sets the
player's prompt colour. If the network connection is closed normally,
without any errors, the colours are reset to default.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>