core: Fix HOLD/WEAR messages when missing body parts + WIELD text
Currently, if you try to wield an item and don't have the required body
parts available the message returned is "You are incapable of wielding
X". For the HOLD and WEAR commands this logic is missing and will
fallthrough causing an incorrect "You can't hold X while holding ." or
"You can't wear X while wearing ." to be displayed.

This commit adds the logic missing from HOLD and WEAR and updates the
text for the WIELD message with the suffix " properly". The messages now
displayed are:

  You are incapable of holding X properly.
  You are incapable of wearing X properly.
  You are incapable of wielding X properly.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>