core: Add OPEN and CLOSE vetoes
Currently the OPEN and CLOSE commands cannot be vetoed. This commit adds
the VetoOpen and VetoClose asKey constants, updates Thing.Unmarshal to
store the vetoes and updates the OPEN and CLOSE commands to check the
vetoes.

The vetoes do not affect the door itself which can still be opened or
closed automatically by registering a trigger event.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>