core: Add additional asKey name constants+setup in Thing.Unmarshal
Currently a Thing has a single name stored in Thing.As[Name]. However it
is desirable to have the name with an initial lower or upper cased
initial letter depending on where it appears in a sentence. It is also
desirable to have a prefix of 'a', 'an' or 'some' in some instance and
'the' in others - also lower or upper cased.

There are two main ways to handling this. We can store Thing.As[Name]
and calculate the upper and lower case versions as well as a 'the'
version where and when needed. This would incur some additional CPU
overhead - potentially for all messages sent. The other option is to
calculate the variants once when a Thing is created and store them as
Thing.As[UName], Thing.As[TheName] and Thing.As[UTheName]. This would
incur some additional memory requirements.

This commit opts for the second approach and stores the additional
values. The implementation is dead simple and easy to revisit in the
future if needed. The Name/TheName variants are lower-cased and
UName/UTheName are the upper-cased variants. The name attribute in the
zone files is expected to be of the forms 'a …', 'an …' or 'some …'.

This commit adds the variants for players in cmd/server which is
currently hard-coded to create players.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>