core: Fix data race in *Thing.Schedule event handling
A data race was introduced in commit a421285 "core: Thing.Schedule
cancel event when it fires to clean up state".

The t.Cancel(event) call is made via time.AfterFunc which does not hold
the state.BWL causing a data race.

This commit removes the t.Cancel(event) from the time.AfterFunc call.
Instead the Cancel call is placed at the start of each specific event
handler which is called with the state.BWL held.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>