core: Fix #DUMP command output multipe spaces collapsing
Due to commit 8f87ce0370d0cfad "cmd,core: Enable folding/wrapping of
text sent to players" consecutive spaces characters are collapsing into
a single space. This causes the tree produced by #DUMP to not render
properly.

This commit updates the #DUMP command to use an intermediate buffer to
capture the output of Thing.Dump and replaces spaces with a hard-space.
WolfMUD uses the symbol for space '␠' U+2420, UTF8 0xe290a0, for a
hard-space which is visible and somewhat obvious.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>