client: Clean up client networking code
Before adding a frontend to the game, this commit cleans up the client
networking code. Some code that is not related to networking has been
moved to a new client/frontend.go source file.

The $POOF command has been updated to no longer include the text "Bye
bye!" as this is now handled in the networking code.

These changes have only had light testing, more comprehensive testing
should be done one the frontend is in place.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>