core: Fix Unmarshal of SPAWN+WAIT attributes of RESET field
Currently when a RESET field is unmarshaled the SPAWN and WAIT
attributes are incorrectly processed. If the SPAWN or WAIT fields are
set then the Thing.Is flags are set irrespective of whether the
attributes are set to true or false.

This commit fixes the unmarshaling of the SPAWN and RESET attributes of
the RESET field and the setting of the respective Thing.Is flags.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>