core: Add constants for player account information to types
Currently there is no provision for storing a player's account
information.

This commit adds constants for:

  Thing.As[Account]  - MD5 hash of the account id/player file
  Thing.As[Password] - Salted SHA512 has of the account password
  Thing.As[Salt]     - The salt used for the account password
  Thing.Int[Created] - Timestamp (Unix Epoc) when account created

Note: that the plain account ID or password is never stored. Only hashes
are stored, if we need to compare account IDs or passwords we calculate
the hash and compare those.

Warning: Currently anyone with access to the server would be able to use
the #DUMP command on a player and see the hashes that have been stored.
This will be delt with in a future commit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>