core,cmd,client: Enable new frontend login+account creation
This commit enables the new frontend for logins and account creation.

The $POOF command has been updated to not display the greeting. The
greeting is now sent to the client via the frontend.

The botrunner has been updated to re-enable the login logic. The login
logic has been updated to check the new format of the login presented by
the new frontend.

Finally, the client code now calls the frontend before entering the
game proper.

NOTE: During account creation a new PLAYER field has been added to the
account header record (the 1st record). The value of PLAYER is the REF
of the player record. Previously we relied on the 2nd record always
being the player record. Having a reference pointing to the player
record is more robust. For old accounts we ifallback to using the 2nd
record and a PLAYER field will be added when saving the player file.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>