core: Add Thing.Marshal and incString helper
This commit adds a Thing.Marshal method for writing Things to a jar
file. This is primarily for writing player files. However, it could also
be used for saving zones - or an entire world.

Things.Marshal uses a small helper Thing.incString that increments a
numeric string by one. For example incString("1") will return the string
"2".

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>