client: Set 'MY' dynamic alias when loading player's inventory
Currently when a player is loaded the dynamic alias "MY" is not set on
the player's immediate inventory items. This means that a player cannot
use the "MY" qualifier when referring to items they load with.

This commit adds the "MY" dynamic qualifier to a player's immediate
inventory items when the player is loaded and reassembled.

NOTE: The "MY" dynamic qualifier is never added to items in containers,
even if the container is in a player's inventory. If the player is
holding an item and puts it into a container they are also holding the
"MY" dynamic qualifier is removed from the item before placing it into
the container.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>