core: Fix saving empty player containers lose container status
When saving a player and their inventory the Inventory record is not
written to the jar file for containers if the container has no items.
This is due to empty Inventory records not being written. The effect is
that any empty containers are no longer containers when the next time
the player logs into the server.

This commit updates Thing.Marshal to always write out the Inventory
record for containers, even if the container is empty.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>