frontend: Clear origins when assembling players (all items unique)
When loading and assembling a player the client.assemblePlayer calls
Thing.InitOnce which will set Thing.Ref[Origin] if Thing.Is&Spawnable is
not set. However, this is incorrect for player items that are being
reloaded and causes all of the player's non-spawnable items to become
unique.

This commit clears Thing.Ref[Origin] after calling Thing.InitOnce as any
unique items the player had should have been removed when the player
quits.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>