core: Fix QUIT/SAVE duplicating unique items
Currently when a player quits their inventory is not checked for unique
items. This enables unique items to be 'duplicated' when the player
saves and reloads with unique items in their inventory. If the player
uses the SAVE command with unique items in their inventory the unique
items are saved in the player file.

This commit adds status.quitUniqueCheck to check for unique items and
force junks them. The check is recursive and will junk unique items in
nested containers.

This commit adds a check for unique items to the command.save function
so they are not included in the player file.

NOTE: If a unique container is forced junked then all of its content
will be force junked as well. The SAVE command will put the content of
unique items into the player file. However, when re-loaded the items
will be restored and then discarded as they are not in any inventory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>