core: Fix zero value lookups using core/type.go maps
Currently it is possible for a lookup as a result of a zero value to
appear to succeed when in fact it fails. For example:

  NameToDir[item.As[Blocker]]

This statement should return the name of the short exit blocked by an
item that is a blocker. For example:

  Door: EXIT→E

item.As[Blocker] would return 'E' and NameToDir['E'] would return the
value of the constant East, which is currently 2.

However, for an exit "N" or "North" the returned constant for North is
currently 0. This is the same as for any failure when using NameToDir
lookups. For example:

   Door: EXIT→Z
  Exits: Z→L1

This commit redefines the zero value constants, where appropriate, as a
"bad key". For example: BadAsKey, BadAnyKey, BadIntKey and BadRefKey. If
a lookup returns one of the bad keys we know the lookup failed.

Note: we can still use "if value, ok := m[key]; ok { }" but it would
mean unwinding nested uses such as: NameToDir[item.As[Blocker]]

This commit also reimplements the refKey.ReverseDir method as a simpler
map lookup. This allows the restriction on where the _North and North
constants can be placed. The only restriction now is that they are kept
in the order given and consecutive. Although even these restrictions may
not be needed now.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>