core,doc: Add support to marshal/unmarshal barriers
This commit adds constants for Barrier (asKey), BarrierAllow (anyKey)
and BarrierDeny (anyKey). These constants are for storing the barrier
attributes in a Thing. This commit also adds code to marshal and
unmarshal a barrier recordjar record for a Thing.

A minor typo in docs/zone-files.txt related to barrier has been
corrected.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>