core: MOVE command now supports barriers, improve barrier help
This commit adds support for checking barriers. Not too happy with the
implementation as it makes the state.Move method more CPU intensive.

This could be handled differently in a number of ways:

  - Maintain a Thing.Is flag and set/clear it based on whether there are
    any barriers or blockers.
  - Keep bookkeeping light-weight and don't allow moveable item
    barriers.
  - Don't let barriers apply to players.

For now, it is what it is :(

This implementation of barriers handles setting only ALLOW or DENY
aliases. Handles players better, and it does not matter if some wise-ass
player is called "Player", "NPC", "Guard", etc.

Barrier documentation in docs/zone-files.txt updated and clarified.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>