core: Switch search order of #DUMP command
Currently the #DUMP command looks at items the actor has and then items
at the location. This can make it difficult to #DUMP an item at a
location where the alias also matches an item the actor has. This often
requires resorting to '#DUMP @' or '#DUMP SELF' and scanning through a
long list.

This commit changes #DUMP to look at items at the location and then
items the actor has. If there is an item matching an alias at the
location and an item the actor has then the 'MY' qualifier can be used
to look at the item the actor has.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>