mailbox: Add a message suffix + restructure mailboxes
Currently the handling of a player's prompt is awkward and handled in
the client package. This makes difficult to access and modify the prompt
from in-game. Accessing and updating the prompt in-game is going to be
necessary for displaying player statistics. There also needs to be a way
of turning off the prompt in order to tidy up player login and logoff.

This commit adds a suffix, per mailbox, that will be appended to the end
of any messages that are subsequently sent. The suffix can be set using
a new Suffix function. The Suffix function can be used anywhere to
safely change the current suffix used.

In order to implement the suffix a new mailbox type has been defined. A
mailbox map then replaces the previous mbox channel and last maps. The
message queue channel and last message sent are now held within the
mailbox type.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>