core,client: Clean-up prompt handling + add frontend/game prompts
Currently the player's prompt is handled by the client.messenger method.
However, this causes an issue as the prompt cannot be easily controlled
from the front-end or while in-game. This causes odd stray prompts to be
displayed when quitting. It also poses a problem when we need to display
the player's statistics in the prompt.

This commit switches to using the new mailbox.Suffix function for
displaying the player's prompt. Prompt handling is split into 3 main
areas:

  front-end: sets prompt on entry and clears on exit
    in-game: $POOF sets the prompt, QUIT clears it
     client: clears prompt when player disconnects

This allows us to easily have different prompts for the frontend and
in-game. A new method state.Prompt has been added for commands to use,
it is planned that this will become the dynamic prompt builder in the
future.

Minor tweak to "Welcome back...", "You step into another world...", "You
leave this world behind." and "Bye bye!" messages for consistent line
spacing. Bug comment removed from state.Poof method.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>