core: Fix OPEN/CLOSE notifications
Currently the OPEN and CLOSE commands do not notify the actor if the
OPEN or CLOSE was successful when the location is crowded. In addition
the crowded check at the actor's location is used to determine whether
notifications are seen on the 'other side'.

This commit corrects the notifications. The actor is now always notified
of the outcome of their action[1]. Observers with the actor are notified
if the actor's location is not crowded. Observers on the other side are
notified if a separate check determines the location is not crowded.

 [1] Unless the actor is also the item being opened/closed because it's
     opening/closing automatically on its own.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>