stats: New server statistics collection and reporting package
This commit adds a new stats package to gather, record and report server
statistics. The statistics are configured via Stats.Rate and Stats.GC in
the server configuration file and written to the server log. The format
of logged statistics is:

  A[n] O[n ±n] T[n ±n] E[n ±n] P[n max]

Where:

  A[    n] - runtime allocations since last collection
  O[n  ±n] - runtime objects / change since last collection
  T[n  ±n] - Thing in the world / change since last collection
  E[n  ±n] - in-flight active events / change since last collection
  P[n max] - current number of players / maximum number of players

The docs/configuration-file.txt has been updated for the new statistics
format.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>