client: Add cfg.debugPanic + panic handler
If a client goroutine panics it should not bring down the whole server.

This commit adds cfg.debugPanic to the package configuration - which is
set via the Debug.Panic setting in the server configuration. A panic
handler has been added to the client.receive method.

If Debug.Panic is set to false (the default) the panic handler will
cause the player to be gracefully removed from the server - player
saved, network connections closed and mailbox removed. A stack trace
will be written to the server log.

If Debug.Panic is set to true then the panic handler has no effect and
the server is left to panic - useful when debugging.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>