core: Add debugThings to pkg cfg + additional Thing logging
This commit adds debugThings, configured via Debug.Things in the server
configuration, to the core package configuration. If Debug.Things is set
to true the server will log when Things are created, freed or finalised.

For example:

  NewThing: 0xc0001ca550 - #UID-44F
  Freed: 0xc0001ca550 - #UID-44F (a sword)
  Finalizing: 0xc0001ca550 - #UID-44F (Freed: true)

These messages can cause a lot of additional noise in the log and should
only be enabled for debugging. If the "Finalizing..." message ends in
"(Freed: false)" this indicates core.Free was not called when disposing
of a item, this may prevent the garbage collector from releasing the
storage.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>