core Make Thing.Schedule/Cancel/Suspend wrappers
This commit makes the Thing.Schedule, Thing.Cancel and Thing.Suspend
wrapper methods to their unexported counterparts. This will allow
logging to be added, however internal calls made by Schedule to cancel
and cancel to suspend will not be logged cutting down some of the noise
generated when logging events.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>