core,doc: New Thing.logEvent helper + documentation updated
The logEvent helper writes the details of an event to the log if event
logging is enabled in the server configuration by setting Debug.Events
to true. Events are logged when they are scheduled, suspended, delivered
or cancelled. For example:

  Event cancel: Reset, for: #UID-9 (a bean), at: -, in: -
  Event firing: Reset, for: #UID-9 (a bean), at: Dec 29 15:55:56, in: 0s
  Event re-schedule: Action, for: #UID-5 (a frog), at: Dec 29 16:10:31, in: 9s
  Event schedule: Reset, for: #UID-9 (a bean), at: Dec 29 15:55:56, in: 1m6.2s

The documentation for Debug.Events in docs/configuration-file.txt has
been updated to reflect the additional logging.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>