core: remove Thing.Event map entry when event cancelled/cleaned-up
Currently when an event is cancelled or cleaned up (after it is
delivered) the map entry in Thing.Event is not removed. This can cause
the event to be copied and triggered - e.g. by spawning. However, in the
case of spawning the reset values are cleared resulting in a NOP.

This commit removes the event map entry in Thing.Event when an event is
cancelled or cleaned up.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>