client: Add logClient to package configuration + new Log method
Currently the server will always log the IP address of an incomming
player connection.

This commit adds logClient, configured via Server.LogClient in the
server configuration, to the client package configuration.

A new client.Log method has been added. It functions like a call to
fmt.Printf but automatically appends the UID uniquily identifying the
connection within the current log. If the configuration setting
Server.LogClient is set to false then the IP address of the incomming
connection will be re-written to '???'. For example:

  [#UID-6] connection from: ???:35848
  [#UID-6] client error: read tcp 127.0.0.1:4001->???:35848: i/o timeout
  [#UID-6] disconnect from: ???:35848

Updated Server.LogClient documentation in docs/configuration-file.txt as
the incoming port is now recorded.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>