core: Add #DEBUG with MEMPROF, CPUPROF and PANIC sub-commands
This commit adds the #DEBUG command with three sub-sommands: MEMPROF,
CPUPROF and PANIC. The availability of the #DEBUG command is configured
via the Debug.AllowDebug configuration setting in the server
configuration file.

The MEMPROF and CPUPROF commands can be used to start and stop memory
and CPU profiling. A profile can be started by supplying ON or START to
the sub-commands. The profiling can then be stopped by supplying OFF,
END or STOP to the sub-commands. The profiles will be written as memprof
or cpuprof into the current directory and can then be inspected using
the pprof tools.

The PANIC sub-command will cause a panic in the state parser and is
intended for testing the client panic handler.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>