core: Fix #GOTO/#TELEPORT command
In commits ac250f7c8ee6449d "core: Refactor Thing.SortedIn→Things.Sort,
new Things type" and c658d7af738a1847 "core,cmd: Move players to
Thing.Who + make commands aware" players were moved from Thing.In to
Thing.Who and associated code updated for the change.

However, the #GOTO/#TELEPORT commands were not updated. This results in
weird behaviour when a player uses the #GOTO/#TELEPORT commands. The
player is moved from Thing.Who to Thing.In, when they next move they are
put back into Thing.Who but the Thing.In reference is not removed. The
result is that a player can appear to be in two places at once.

This commit updates the Thing.In references to be Thing.Who in the
state.Teleport method used by #GOTO/#TELEPORT to fix the issue.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>