core: Update state.Prompt to take player and fmt.Printf like args
This commit updates the state.Prompt method to take the player whose
prompt should be set - instead of always setting s.actor's prompt.
Instead a taking a simple text string for the prompt the method now
takes a format string and arguments - as per a fmt.Printf call.

The QUIT and POOF commands have been updated for the changes.

NOTE: As we are not using the *state within the method now would it be
better to move the method to Thing?

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>