core: Make all message sent to a specific player priority
At the moment only messages sent to the actor are sent priority.

However, if there is a participant involved then messages to the
participant will be de-spammed and de-duplicated - resulting in the
participant only being notified the first time something happens.

For example, a HIT command is being implemented in order to test player
health for the upcoming addition of combat. If the participant is hit
twice in succession, with no other intervening messages, they only
receive one message indicating they have been hit.

This commit makes all messages sent to a specific player priority so
that they are not de-spammed or de-duplicated. Currently this only
impacts the EXAMINE command. When a player examines another player the
participant (other player) is notified with a "<actor> studies you."
message. Previously the participant would have seen one message if the
actor examined them repeatedly, now the observer is notified each time.

An alternative solution would be to add an observer *Thing reference to
the state struct and modify the check in mailman to be 'ref == s.actor
|| ref == s.observer' and set s.observer = nil at the end of mailman.
This would solve the EXAMINE spam issue and fix the issue with losing
messages for the HIT command, but at the cost of increased memory usage.
As a participant is not used very often is the additional memory
overhead worth it? Maybe revisit this during the big code clean-up...

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>