client: Take BWL before disposing of player (data race)
In client.clean the player Thing is accessed while disposing of the
player. However, the player has been in the game world and manipulated
while being under the BRL. In client.clean the player is not under the
BRL which leads to a data race.

This commit takes the BWL before disposing of the player Thing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>